cleanup distributed arrayvec fusion
